+++ include/dev/world/chunk.h
@@ -3,10 +3,12 @@
 
 
 #include <vector>
-#include <block.h>
 #include <glad/glad.h>
 #include <glm/glm.hpp>
 
+#include <block.h>
+#include <hash.h>
+#include <voxelVertex.h>
 
 #define CHUNKSIZE 15
 
@@ -15,35 +17,8 @@ struct chunkBuffers {
 };
 
 
-struct Vec3Hash {
-	std::size_t operator()(const glm::ivec3& v) const {
-		std::size_t h1 = std::hash<int>()(v.x);
-		std::size_t h2 = std::hash<int>()(v.y);
-		std::size_t h3 = std::hash<int>()(v.z);
-		return h1 ^ (h2 << 1) ^ (h3 << 2); // Combinação simples
-	}
-};
 
-struct PairHash {
-	std::size_t operator()(const std::pair<int, int>& p) const {
-		return std::hash<int>()(p.first) ^ (std::hash<int>()(p.second) << 1);
-	}
-};
 
-struct Vertex
-{
-	char X, Y, Z;
-	float uvX, uvY;
-	char normalX, normalY, normalZ;
-	uint8_t lightLevel;
-
-	Vertex(char _posX, char _posY, char _posZ, float _texGridX, float _texGridY, char _normalX, char _normalY, char _normalZ, uint8_t _lightLevel)
-		: X(_posX), Y(_posY), Z(_posZ), 
-		uvX(_texGridX), uvY(_texGridY), 
-		normalX(_normalX), normalY(_normalY), normalZ(_normalZ),
-		lightLevel(_lightLevel){
-	}
-};
 
 class Chunk {
 public:
@@ -81,7 +56,7 @@ public:
 	std::vector<Block> populateChunk(glm::ivec3 chunkCoords);
 	void addVertxInfo(FACE face, char x, char y, char z, std::vector<Vertex>& vertices, std::vector<unsigned int>& indices, unsigned int& currentVertex, Block storedBlock);
 	void regenMesh();
-	void calculateChunkLighting(std::unordered_map<glm::ivec3, Chunk, Vec3Hash>& WorldData, std::unordered_map<std::pair<int, int>, int, PairHash> &highestChunkY);
+
 
 	bool isChunkEmpty() {
